decoupling parse/resolve : rfct
